图片文件缓存优化，提升效率

+++ AndBase/src/com/ab/download/AbDownloadProgressListener.java
@@ -30,7 +30,7 @@ public interface AbDownloadProgressListener {
 	 *
 	 * @param size the size
 	 */
-	public void onDownloadSize(int size);
+	public void onDownloadSize(long size);
 }
